added location devise gem and updated models

+++ db/migrate/20151212233033_create_firework_details.rb
@@ -1,7 +1,7 @@
 class CreateFireworkDetails < ActiveRecord::Migration
   def change
     create_table :firework_details do |t|
-      t.integer :cost
+      t.decimal :cost
       t.integer :purchased
       t.belongs_to :firework, index: true
